requirements.txt
Django ~= 4.2.3
celery >= 5.2.7, < 6
django-bootstrap-form >= 3.4, < 4
oauthlib >= 3.2.2, < 4
requests >= 2.28.2, < 3
git+https://github.com/apluslms/django-lti-login.git@3.3.0#egg=django-lti-login==3.3.0
git+https://github.com/apluslms/django-essentials.git@1.6.0#egg=django-essentials==1.6.0
git+https://github.com/apluslms/a-plus-client.git@v1.3.0#egg=a_plus_client
git+https://github.com/apluslms/greedy-string-tiling.git@v0.13.0
aplus-auth ~= 0.2.3
setuptools
django-debug-toolbar
pytz
pylibmc
pytest-playwright
pytest-django
pytest-env
